VASTLY simplify and improve logic for storing entries
Merged everything into only one EntryList, it is now
trivial to get indices and render them properly. This
also reduces overall memory needed in large directories
by up to 8% on average :D

The only downside is the temporary memory overhead by
having to store the files and append them at the end to
the main EntryList to keep the order of directories
first, then files.