perf: Up to 390,000x performance when creating selections on...
...large directories (tested on directories with
50k+ files)

Time before: ~12.3s
Time after: ~30.5us
Rough performance increase: 390,000x

The compiler can now trivially vectorize the
and optimize instructions to invert the selection.

But there is a disadvantage, now the program has
to keep state of whether each entry is part of the
selection or not, increasing memory footprint.